Using default parameters when [ --config "" ] is used. Updated default: sendNoObjDetectedEvents = true. fixed mirror view when autoScale is set on/off
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@397 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ src/FindObject.cpp
@@ -441,7 +441,7 @@ protected:
 			unsigned int threadCounts = Settings::getGeneral_threads();
 			if(threadCounts == 0)
 			{
-				threadCounts = tilts.size();
+				threadCounts = (unsigned int)tilts.size();
 			}
 
 			for(unsigned int i=0; i<tilts.size(); i+=threadCounts)
